[style] change height table

+++ Siswa/keranjang.php
@@ -1,7 +1,7 @@
 <?php
 include "header.php";
 ?>
-<div class="container" style="min-height: 87vh; padding-top: 5rem;">
+<div class="container" style="min-height: 100vh; padding-top: 5rem;">
     <h2 class="text-center pb-3">Daftar Buku di Keranjang</h2>
     <table class="table table-hover striped">
         <thead>
@@ -14,25 +14,32 @@ include "header.php";
         </thead>
         <tbody>
             <?php
-            foreach (@$_SESSION['cart'] as $key_produk => $val_produk): ?>
-                <tr>
-                    <td>
-                        <?= ($key_produk + 1) ?>
-                    </td>
-                    <td>
-                        <?= $val_produk['nama_buku'] ?>
-                    </td>
-                    <td>
-                        <?= $val_produk['qty'] ?>
-                    </td>
-                    <td><a href="hapus_cart.php?id=<?= $key_produk ?>" class="btn btn-danger"><strong>X</strong></a></td>
-                </tr>
+            if (isset($_SESSION['cart']) && !empty($_SESSION['cart'])) {
+                foreach (@$_SESSION['cart'] as $key_produk => $val_produk):
+                    ?>
 
-            <?php endforeach ?>
+                    <tr>
+                        <td>
+                            <?= ($key_produk + 1) ?>
+                        </td>
+                        <td>
+                            <?= $val_produk['nama_buku'] ?>
+                        </td>
+                        <td>
+                            <?= $val_produk['qty'] ?>
+                        </td>
+                        <td><a href="hapus_cart.php?id=<?= $key_produk ?>" class="btn btn-danger"><strong>X</strong></a></td>
+                    </tr>
+
+                <?php endforeach;
+            } else {
+                echo '<tr><td colspan="12">Keranjang Anda kosong.</td></tr>';
+            }
+            ?>
         </tbody>
     </table>
-    <a href="buku.php" class="btn btn-secondary">Pinjam Lagi</a>
-    <a href="checkout.php" class="btn btn-primary">Check Out</a>
+    <a href="buku.php" class="btn btn-primary">Pinjam Lagi</a>
+    <a href="checkout.php" class="btn btn-danger">Check Out</a>
 </div>
 <?php
 include "footer.php";
